+++ src/lib.rs
@@ -19,3 +19,5 @@ pub use killstreaker::Killstreaker;
 pub use paint::Paint;
 pub use strange_part::StrangePart;
 pub use wear::Wear;
+
+pub use strum;